@@ -21,82 +21,84 @@
     "release:github": "semantic-release --extends './scripts/release/releaserc.github.js'",
     "release:information": "semantic-release --dry-run --extends './scripts/release/releaserc.information.js' --no-ci",
     "repair": "npx rimraf build node_modules .git/hooks package-lock.json && npm install && npx husky install",
-    "start": "react-scripts start",
+    "start": "vite",
     "start:prod": "npx http-server ./build --proxy http://localhost:8080? --silent",
     "stylelint-check": "stylelint-config-prettier-check",
-    "test": "react-scripts test"
+    "test": "react-scripts test --transformIgnorePatterns 'node_modules/(?!(@ionic/react|@ionic/react-router|@ionic/core|@stencil/core|ionicons)/)'",
+    "serve": "vite preview",
+    "prepare": "husky install"
   },
   "dependencies": {
-    "@capacitor/android": "^3.0.0-beta.1",
-    "@capacitor/core": "^3.0.0-beta.1",
-    "@capacitor/filesystem": "^0.5.6",
-    "@capacitor/ios": "^3.0.0-beta.1",
-    "@capacitor/splash-screen": "^0.3.10",
-    "@capacitor/status-bar": "^0.4.10",
-    "@capacitor/storage": "^0.3.10",
-    "@ionic/react": "^5.6.6",
-    "@ionic/react-router": "^5.6.6",
-    "@ionic/storage": "^3.0.4",
-    "@reduxjs/toolkit": "^1.5.1",
-    "@testing-library/jest-dom": "5.12.0",
-    "@testing-library/react": "^11.2.6",
-    "@testing-library/user-event": "^13.1.9",
-    "fuse.js": "^6.4.6",
-    "i18next": "^20.2.2",
-    "i18next-browser-languagedetector": "6.1.1",
-    "ionicons": "5.5.1",
+    "@capacitor/android": "5.0.0",
+    "@capacitor/core": "^5.0.0",
+    "@capacitor/filesystem": "^5.0.0",
+    "@capacitor/ios": "5.2.3",
+    "@capacitor/preferences": "^5.0.0",
+    "@capacitor/splash-screen": "^5.0.0",
+    "@capacitor/status-bar": "^5.0.0",
+    "@ionic/react": "^7.2.1",
+    "@ionic/react-router": "^7.2.1",
+    "@ionic/storage": "4.0.0",
+    "@reduxjs/toolkit": "1.9.5",
+    "app-provider": "^1.1.2",
+    "fuse.js": "6.6.2",
+    "i18next": "23.4.4",
+    "i18next-browser-languagedetector": "7.1.0",
+    "ionicons": "7.1.2",
     "md5": "^2.3.0",
     "microsoft-cognitiveservices-speech-sdk": "^1.16.0",
-    "query-string": "^7.0.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-i18next": "^11.8.15",
-    "react-redux": "7.2.4",
-    "react-router": "5.2.0",
-    "react-router-dom": "5.2.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.1.0",
-    "redux-persist": "^6.0.0",
-    "typescript": "^4.2.4",
-    "use-query-params": "^1.2.2"
+    "query-string": "8.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-i18next": "13.1.2",
+    "react-redux": "8.1.2",
+    "react-router": "5.3.4",
+    "react-router-dom": "5.3.4",
+    "redux-persist": "6.0.0",
+    "sharp": "^0.32.4",
+    "swiper": "^10.1.0",
+    "typescript": "5.1.6",
+    "use-query-params": "2.2.1"
   },
   "devDependencies": {
-    "@capacitor/cli": "^3.0.0-beta.1",
-    "@commitlint/cli": "^12.1.4",
-    "@commitlint/config-conventional": "^12.1.4",
-    "@ionic/cli": "^6.15.0",
-    "@semantic-release/changelog": "^5.0.1",
-    "@semantic-release/exec": "^5.0.0",
-    "@semantic-release/git": "^9.0.0",
-    "@squoosh/cli": "^0.6.4",
-    "@types/jest": "^26.0.23",
+    "@boehringer-ingelheim/eslint-config": "1.1.4",
+    "@boehringer-ingelheim/prettier-config": "1.0.0",
+    "@capacitor/cli": "^5.0.0",
+    "@commitlint/cli": "17.6.7",
+    "@commitlint/config-conventional": "17.6.7",
+    "@ionic/cli": "7.1.1",
+    "@playwright/test": "^1.37.0",
+    "@semantic-release/changelog": "6.0.3",
+    "@semantic-release/exec": "6.0.3",
+    "@semantic-release/git": "10.0.1",
+    "@types/jest": "^26.0.24",
     "@types/lunr": "^2.3.3",
-    "@types/node": "^15.0.3",
-    "@types/react": "^17.0.5",
-    "@types/react-dom": "^17.0.5",
+    "@types/node": "20.5.0",
+    "@types/react": "18.0.37",
+    "@types/react-dom": "18.0.11",
     "@types/react-redux": "^7.1.16",
-    "@types/react-router": "^5.1.14",
-    "@types/react-router-dom": "5.1.7",
-    "chalk": "^4.1.1",
+    "@types/react-router": "5.1.20",
+    "@types/react-router-dom": "5.3.3",
+    "@vitejs/plugin-react": "4.0.3",
+    "chalk": "5.3.0",
     "device-bars": "^1.1.2",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-jest-playwright": "0.3.3",
-    "eslint-plugin-prettier": "^3.4.0",
-    "husky": "^6.0.0",
-    "jest-playwright-preset": "^1.5.2",
-    "lint-staged": "^11.0.0",
+    "eslint-config-prettier": "9.0.0",
+    "eslint-plugin-jest-playwright": "0.9.0",
+    "eslint-plugin-prettier": "5.0.0",
+    "husky": "8.0.3",
+    "jest-playwright-preset": "3.0.1",
+    "lint-staged": "14.0.0",
     "mobicon": "3.0.0",
     "mobisplash": "0.2.2",
-    "ora": "^5.4.0",
-    "playwright": "1.11.0",
-    "prettier": "^2.3.0",
-    "semantic-release": "^17.4.3",
-    "stylelint": "^13.13.1",
-    "stylelint-config-prettier": "^8.0.2",
-    "stylelint-config-standard": "^22.0.0",
-    "stylelint-order": "^4.1.0",
-    "stylelint-prettier": "^1.2.0",
-    "ts-jest": "^26.5.6"
+    "prettier": "3.0.1",
+    "react-scripts": "5.0.1",
+    "semantic-release": "21.0.7",
+    "stylelint": "^15.10.2",
+    "stylelint-config-standard": "34.0.0",
+    "stylelint-order": "6.0.3",
+    "stylelint-prettier": "4.0.2",
+    "ts-jest": "^26.5.6",
+    "vite": "^4.4.9"
   },
   "lint-staged": {
     "src/**/*.{js,jsx,ts,tsx,json,css,scss,md}": [
